Splitting centerPanel.svelte into independent components
centerPanel.svelte can be decomposed as headerPanel.svelte, chartPanel.svelte, and orderPanel.svelte. This approach gives more consistent and manageable approaches. Plus, used Svelte store to keep which market the user is watching, to share the same data across all Svelte pages in /exchange endpoint.
