+++ frontend/src/routes/exchange/+page.svelte
@@ -4,8 +4,12 @@
     import { onMount, onDestroy } from 'svelte';
     import { getAllAvailableMarketsInfo } from './functions/getMarketInfo.js';
     import { balances } from '../../stores/usesrAssets';
-    import CenterPanel from './centerPanel.svelte';
+    import { selectedMarketCode } from './stores/selectedMarketStore.js';
+    import HeaderPanel from './headerPanel.svelte';
     import LeftPanel from './leftPanel.svelte';
+    import ChartPanel from './chartPanel.svelte';
+    import OrderPanel from './orderPanel.svelte';
+    import { get } from 'svelte/store';
 
     let marketData = [];             // Market data
     let selectedMarket = null;       // Selected market (the market that the user is currently viewing)
@@ -17,11 +21,12 @@
     let sortKey = '';                // The key to sort the market data by (e.g., market, signed_change_rate, trade_price, acc_trade_price_24h)
     let sortOrder = 'asc';           // The sort order (asc for ascending, desc for descending)
 
-    // Function to fetch market data and update the state
     async function fetchMarketData() {
         try {
             marketData = await getAllAvailableMarketsInfo();
-            selectedMarket = selectedMarket || marketData[0];
+            const initialMarketCode = get(selectedMarketCode);  // Get the initial value from the store
+            selectedMarket = marketData.find(market => market.market === initialMarketCode) || marketData[0];
+            selectedMarketCode.set(selectedMarket.market);
             selectedMarketCodeUnit = selectedMarket.market.split('-')[1].toUpperCase();
             currentPrice = selectedMarket.trade_price;
             sortMarketData();
@@ -32,26 +37,22 @@
 
     onMount(async () => {
         await fetchMarketData();
-
-        // Fetch the market data every 2 seconds
         intervalId = setInterval(fetchMarketData, 2000);
     });
 
-    // Clear the interval when the component is destroyed
     onDestroy(() => {
         if (intervalId) {
             clearInterval(intervalId);
         }
     });
 
-    // Set the selected market
     function setSelectedMarket(market) {
         selectedMarket = market;
+        selectedMarketCode.set(market.market);
         selectedMarketCodeUnit = market.market.split('-')[1].toUpperCase();
         currentPrice = market.trade_price;
     }
 
-    // Handle the sorting of the market data
     function handleSort(event) {
         const { key } = event.detail;
         if (sortKey === key) {
@@ -63,7 +64,6 @@
         sortMarketData();
     }
 
-    // Sort the market data based on the sort key and sort order
     function sortMarketData() {
         if (!sortKey) return;
 
@@ -89,13 +89,15 @@
     });
 </script>
 
-<main class="h-screen bg-gray-100 font-sans overflow-hidden">
+<main class="h-[130vh] font-sans overflow-hidden">
     <div class="grid grid-cols-10 h-full">
-        <div class="col-span-3 h-full overflow-y-auto">
+        <div class="col-span-3 h-full">
             <LeftPanel {marketData} {setSelectedMarket} {selectedMarket} on:sort={handleSort} />
         </div>
-        <div class="col-span-7 h-full overflow-y-auto">
-            <CenterPanel {selectedMarket} {availableBalance} {currentPrice} />
+        <div class="col-span-7 h-full">
+            <HeaderPanel {selectedMarket} {selectedMarketCodeUnit} />
+            <ChartPanel {selectedMarket} {selectedMarketCodeUnit} />
+            <OrderPanel {selectedMarket} {availableBalance} {currentPrice} />
         </div>
     </div>
 </main>
@@ -108,6 +110,8 @@
         overflow: hidden;
     }
     .col-span-3, .col-span-7 {
-        overflow-y: auto;
+        height: 100%;
+        display: flex;
+        flex-direction: column;
     }
 </style>
